+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # PHP-Enum Doctrine changelog
 
+## Unreleased
+
+* **[CHANGED]** Require `zlikavac32/php-enum:^1.3`
+
 ## 0.1.0 (2017-11-17)
 
 * **[NEW]** First tagged version